refactor(config): extract sync bundle logic from sync-service.ts (867→783 lines)
Extract readIndexFile, bundleSyncUnit, collectAllSyncUnits into
sync-bundle.ts. These form a cohesive "bundle creation" concern.
Backward-compatible re-exports preserved.
